elf: fix potential overflow when reusing TextBlock
If `TextBlock` does not have a successor, its capacity in the worst
case will equal to almost the entire `u64` resolution, meaning
scaling it by `alloc_num / alloc_den` will cause an overflow and a
panic in debug/release-safe mode.

This fix is backported from the `MachO` change in c1768d9.