prebuild.zig
This is an idea to add support for a prebuild.zig file alongside build.zig.  prebuild.zig is a normal zig program with a "main" function.  Its always executed before "build.zig".  This gives projects the opportunity to hook into the Zig build system before their build.zig code is compiled.  This means it can be used to prepare anything required in order to get build.zig to compile.  The main use case for this is to prepare dependencies that build.zig requires.