force buildpkgs.has to be called at comptime
The reason we want to force buildpkgs.has to be comptime is it is meant to be surrounding a block that contains an @import, so if a user accidently forgets to mark it as comptime then their build.zig will be invalid, but this error will only be caught with certain configurations.  I've marked @import("buildpkgs").has with callconv(.Inline) which should force it to be evaluated at comptime, but it looks like this behavior is not implemented yet.