stage2: constant data in .rodata section instead of .text
For each section containing decls, a `TextBlockList` is created in
`Elf`, containing the available blocks in this section, as well as the
phdr and section indexes. All functions manipulating `TextBlock` now
take a `TextBlockList` indicating where the block should be stored. The
`TextBlockList` to use is determined by `getDeclBlockList()`.

Previously all decl were saved in the .text section of the ELF.
Now constant data is put in the .rodata section of the ELF.