Fix AtomicCondition compile error ziglang/zig#8051
This also changes the `wait(...)` signature to take a `Mutex.Impl.Held`
object instead of a plain `Mutex`, which IIUC is clearer to the intent
of the function (ie., you should be passing in a locked mutex, not just
any mutex).

For example:

    var mutex = Mutex{};
    var cond = Condition{};

    var held = mutex.acquire();
    defer held.release();

    while (bad_state == true) {
        // cond.wait(&mutex); // <-- Original
        cond.wait(&held);     // <-- Updated
    }

Not really a breaking change for code using `AtomicCondition` (since it
didn't compile before), but this does affect the other implementations
and would presumably break anything using that code as well.